+++ 01_Docker源码/ESurfingDialer-Docker-Fixed/src/main/kotlin/com/rsplwe/esurfing/network/NetResult.kt
@@ -3,7 +3,7 @@ package com.rsplwe.esurfing.network
 sealed class NetResult<out T : Any> {
 
     data class Success<out T : Any>(val data: T) : NetResult<T>()
-    data class Error(val exception: String) : NetResult<Nothing>()
+    data class Error(val exception: Throwable) : NetResult<Nothing>()
 
     override fun toString(): String {
         return when (this) {